% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.R
\name{createSBGNviewObject}
\alias{createSBGNviewObject}
\title{SBGNview S3 class}
\format{
A list containing two elements:
          1. data: 
             A list contains information for all input SBGN pathways. Each element is a list (e.g. 'P00001') holding information of a single SBGN pathway. The information include parsed information of glyphs, arcs and SBGNview parameters passed to SBGNview function
             
          2. output.file:
             A string of the path to the output file. It is the string set by parameter 'output.file' in function SBGNview.
}
\usage{
createSBGNviewObject(data, output.file)
}
\description{
This S3 class of objects contains information generated by the SBGNview function. A demo is available by running 'data(SBGNview.obj)'.
The SBGNview object can be modified by using the "+" binary operator in conjunction with the \code{\link{highlightArcs}}, \code{\link{highlightNodes}}, and \code{\link{highlightPath}} functions. 
This mechanism to allow layer-by-layer graph modification is similar to that of ggplot2, a widely used R package for data visualization. 
Defining this class as a formal (S4) class will produce an S4 SBGNview object which doesn't work with the binary operator and doesn't allow for layer-by-layer graph modification. 
Therefore, we decided to go with the S3 implementation of the SBGNview class since it works well with the binary operator making this 
functionality more intuitive and user friendly.
}
\examples{
data(SBGNview.obj)
}
