% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.utilities.R
\name{downloadSbgnFile}
\alias{downloadSbgnFile}
\title{Download pre-generated SBGN-ML file from GitHub}
\usage{
downloadSbgnFile(pathway.id, download.folder = ".")
}
\arguments{
\item{pathway.id}{The ID of pathway. It is a pathway database specific ID. For accepted pathway IDs, please check \code{data("pathways.info")}. IDs are in column "pathway.id" (pathways.info[,"pathway.id"])}

\item{download.folder}{The output folder to store downloaded SBGN-ML files.}
}
\value{
A character string. The path to the downloaded SBGN-ML file.
}
\description{
This function can download our pre-generated SBGN-ML files from \href{https://github.com/datapplab/SBGNhub/tree/master/data/SBGN}{GitHub repository}.
}
\examples{
data("pathways.info")
data(sbgn.xmls)
input.sbgn = downloadSbgnFile(
                  pathway.id = pathways.info[1,"pathway.id"],
                  download.folder = "./")
}
