% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.obj.fun.R
\name{fileOutput}
\alias{fileOutput}
\title{Retrieve output file information from a SBGNview object}
\usage{
fileOutput(obj)
}
\arguments{
\item{obj}{An SBGNview class object.}
}
\value{
A string. The output file information.
}
\description{
Retrieve output file information from a SBGNview object
}
\details{
This function prints the output file path recorded in a SBGNview object.
}
\examples{
data('SBGNview.obj' )
fileOutput(SBGNview.obj) 
}
