% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.utilities.R
\name{highlight.arcs}
\alias{highlight.arcs}
\title{Highlight arcs by arc class}
\usage{
highlight.arcs(class = "all", color = "black", line.width = 2,
  tip.size = 6)
}
\arguments{
\item{class}{String. The arc class to modify.}

\item{color}{String.}

\item{line.width}{Numeric.}

\item{tip.size}{Numeric}
}
\value{
A function that modify a SBGNview object to change arc color.
}
\description{
This function creates another function that can modify a SBGNview object and modify arc. Normally we use it as an argument to \code{\link{+.SBGNview}} and modify a SBGNview object
}
\examples{
 data(SBGNview.obj)
\dontrun{
data("SBGNview.obj" )
obj.new = SBGNview.obj + 
            highlight.arcs(class = "production",color = "red") 
            
}
}
