% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.utilities.R
\name{highlightArcs}
\alias{highlightArcs}
\title{Highlight arcs by arc class}
\usage{
highlightArcs(class = "all", color = "black", line.width = 2,
  tip.size = 6)
}
\arguments{
\item{class}{String. The arc class to modify.}

\item{color}{String. The color of arcs with selected 'class'}

\item{line.width}{Numeric. Line thickness}

\item{tip.size}{Numeric}
}
\value{
A SBGNview object
}
\description{
This function can modify a SBGNview object's arc. Normally we use it as an argument to \code{\link{+.SBGNview}} and modify a SBGNview object
}
\examples{
data(SBGNview.obj)
obj.new = SBGNview.obj + highlightArcs(class = 'production',color = 'red') 
            
}
