% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.utilities.R
\name{highlightNodes}
\alias{highlightNodes}
\title{Highlight input nodes}
\usage{
highlightNodes(node.set = "all", node.set.id.type = "CompoundName",
  glyphs.id.type = "pathwayCommons", cpd.or.gene = "compound",
  stroke.color = "black", stroke.width = 10, stroke.opacity = 1,
  show.glyph.id = FALSE, select.glyph.class = c(), label.x.shift = 0,
  label.y.shift = 0, label.color = "black", label.font.size = NULL,
  label.spliting.string = NULL, labels = NULL)
}
\arguments{
\item{node.set}{A vector of character strings. Input molecule IDs whose nodes are to be highlighted. It can be any ID types supported by SBGNview.}

\item{node.set.id.type}{A character string. ID type of input nodes.}

\item{glyphs.id.type}{A character string. ID type of nodes in SBGN-ML file.}

\item{cpd.or.gene}{A character string. One of 'gene' or 'compound'.}

\item{stroke.color}{A character string. The border color of highlighted nodes.}

\item{stroke.width}{A character string. The border stroke width of highlighted nodes.}

\item{stroke.opacity}{Numeric. The border stroke opacity of highlighted nodes.}

\item{show.glyph.id}{Logic. If set to 'TRUE', glyph IDs are plotted instead of their labels.}

\item{select.glyph.class}{Character. Select glyphs by class. It should be one of the values of XML attribute 'class' of a 'glyph' element. For example 'macromolecule', "simple chemical"}

\item{label.x.shift}{Numeric. Move labels horizontally.}

\item{label.y.shift}{Numeric}

\item{label.color}{Character}

\item{label.font.size}{Numeric}

\item{label.spliting.string}{Character}

\item{labels}{A vector of character strings. The labels to display on the nodes of input molecules. The names of this vector should be the vector of 'node.set'. The values are the new labels to display.}
}
\value{
A SBGNview obj
}
\description{
Change node properties such as border color and width to highlight a list of input nodes. This function should be used as the second argument to function \code{\link{+.SBGNview}}.
}
\examples{
data(SBGNview.obj)
obj.new = SBGNview.obj + 
       highlightNodes(node.set = c('tyrosine', '(+-)-epinephrine'),
                      stroke.width = 4, stroke.color = 'green') 
}
