% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.utilities.R
\name{loadMappingTable}
\alias{loadMappingTable}
\title{Generate ID mapping table from input and output ID types}
\usage{
loadMappingTable(
  input.type,
  output.type,
  species = NULL,
  cpd.or.gene,
  limit.to.ids = NULL,
  SBGNview.data.folder = "./SBGNview.tmp.data",
  existing.data.folder = NULL
)
}
\arguments{
\item{input.type}{A character string. Gene or compound ID type}

\item{output.type}{A character string. Gene or compound ID type}

\item{species}{A character string. Three letter KEGG species code.}

\item{cpd.or.gene}{A character string. Either 'gene' or 'compound'}

\item{limit.to.ids}{Vector. Molecule IDs of 'input.type'.}

\item{SBGNview.data.folder}{A character string.}
}
\value{
A list containing the mapping table.
}
\description{
This function generates the ID mapping table from input and output ID types. If provided a vector of input IDs (limit.to.ids argument), the function will output mapping table only containing the input IDs. Otherwise, the function will output all IDs of input and output types (restricted to a species if working on gene types and specified  the 'species' parameter).
}
\examples{
 data(mapped.ids)
 entrez.to.pathwayCommons = loadMappingTable(
                                input.type = 'ENTREZID'
                                ,output.type = 'pathwayCommons'
                                ,species = 'hsa'
                                ,cpd.or.gene = 'gene'
                             )
                             
}
