% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.obj.fun.R
\name{output.file<-}
\alias{output.file<-}
\title{Set output file information for a SBGNview object}
\usage{
output.file(obj) <- value
}
\arguments{
\item{obj}{No need to provide}

\item{value}{No need to provide}
}
\value{
A SBGNview object
}
\description{
Set output file information for a SBGNview object
}
\examples{
data("SBGNview.obj" )
\dontrun{
output.file(SBGNview.obj) = "./test.output"
}
}
