% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.obj.fun.R
\name{output.file}
\alias{output.file}
\title{Retrieve output file information from a SBGNview object}
\usage{
output.file(obj)
}
\arguments{
\item{obj}{A SBGNview object.}
}
\value{
A string. The output file information. The same as parameter "output.file" when running \code{\link{SBGNview}}
}
\description{
Retrieve output file information from a SBGNview object
}
\examples{
data("SBGNview.obj" )
\dontrun{
output.file(SBGNview.obj) 
}
}
