% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.obj.fun.R
\name{outputFile<-}
\alias{outputFile<-}
\title{Set output file information for a SBGNview object}
\usage{
outputFile(obj) <- value
}
\arguments{
\item{obj}{A SBGNview class object}

\item{value}{No need to provide}
}
\value{
A SBGNview class object
}
\description{
Set output file information for a SBGNview object
}
\details{
This function sets the output file path recorded in a SBGNview object.
}
\examples{
\dontrun{
outputFile(SBGNview.obj) <- "test.output.file"
}
}
