% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.obj.fun.R
\name{outputFile}
\alias{outputFile}
\title{Retrieve output file information from a SBGNview object}
\usage{
outputFile(obj)
}
\arguments{
\item{obj}{A SBGNview object.}
}
\value{
A string. The output file information. The same as parameter "output.file" when running \code{\link{SBGNview}}
}
\description{
Retrieve output file information from a SBGNview object
}
\details{
This function prints the output file path recorded in a SBGNview object. When printing the SBGNview object, output image files will be generated using the output file.
}
\examples{
data("SBGNview.obj" )
outputFile(SBGNview.obj) 
}
