% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways.info}
\alias{pathways.info}
\title{Information of collected pathways}
\format{A data.frame}
\usage{
pathways.info
}
\description{
Information of collected pathways
}
\keyword{datasets}
