% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways.stats}
\alias{pathways.stats}
\title{Number of pathways collected}
\format{
A data.frame
}
\usage{
data(pathways.stats)
}
\description{
Number of pathways collected
}
\details{
The number of pathways in SBGNhub from each source database. It is calculated from data 'pathways.info'.
}
\keyword{datasets}
