% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.obj.fun.R
\name{+.SBGNview}
\alias{+.SBGNview}
\title{Generic function to modify SBGN graph features}
\usage{
\method{+}{SBGNview}(SBGNview.obj, fn)
}
\arguments{
\item{SBGNview.obj}{An object of class SBGNview}

\item{fn}{A character string. The name of any funtion that modifies and returns a SBGNview object. Some functions are available in SBGNview package: \code{\link{highlightPath}}, \code{\link{highlightNodes}}.}
}
\value{
The returned value of *fn*
}
\description{
This binary operator ('+') modifies a SBGNview object(first argument) using a function (second argument)
}
\examples{
\dontrun{
obj.new <- SBGNview.obj + highlightArcs(class = 'production',
                                        color = 'red') 
}
}
