% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.obj.fun.R
\name{print.SBGNview}
\alias{print.SBGNview}
\title{A wrapper to run function \code{\link{render.sbgn}} for all pathways in a SBGNview object and generate image files.}
\usage{
\method{print}{SBGNview}(x, ...)
}
\arguments{
\item{x}{An object of class SBGNview}

\item{...}{Other parameters passed to print.}
}
\value{
None
}
\description{
A wrapper to run function \code{\link{render.sbgn}} for all pathways in a SBGNview object and generate image files.
}
\examples{
data("SBGNview.obj" )
\dontrun{
print(SBGNview.obj)
}
}
