% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SBGNhub.id.mapping.tables}
\alias{SBGNhub.id.mapping.tables}
\title{Mapping tables available in SBGNhub}
\format{
A matrix
}
\usage{
data(SBGNhub.id.mapping.tables)
}
\description{
Mapping tables available in SBGNhub
}
\details{
A collection of ID mapping table files available in SBGNhub. 
If a matching mapping table is available, it will be downloaded from \href{https://github.com/datapplab/SBGNhub/tree/master/data/id.mapping.unique.pair.name}{SBGNhub} github page.
}
\keyword{datasets}
