% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SBGNview.obj}
\alias{SBGNview.obj}
\title{Demo SBGNview object}
\format{
A SBGNview S3 class object
}
\usage{
SBGNview.obj
}
\description{
Demo SBGNview object
}
\details{
An example SBGNview object for demo purposes. 
This S3 class of objects contains information generated by the SBGNview function.
The SBGNview object can be modified by using the "+" binary operator in conjunction with the \code{\link{highlightArcs}}, \code{\link{highlightNodes}}, and \code{\link{highlightPath}} functions. 
This mechanism to allow layer-by-layer graph modification is similar to that of ggplot2, a widely used R package for data visualization. 
Defining this class as a formal (S4) class will produce an S4 SBGNview object which doesn't work with the binary operator and doesn't allow for layer-by-layer graph modification. 
Therefore, we decided to go with the S3 implementation of the SBGNview class since it works well with the binary operator making this 
functionality more intuitive and user friendly.
}
\keyword{datasets}
