% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.utilities.R
\name{downloadSbgnFile}
\alias{downloadSbgnFile}
\title{Download pre-generated SBGN-ML file from GitHub}
\usage{
downloadSbgnFile(pathway.id, download.folder = ".")
}
\arguments{
\item{pathway.id}{The ID of pathway. For accepted pathway IDs, please check \code{data('pathways.info')}. IDs are in column 'pathway.id' (pathways.info[,'pathway.id'])}

\item{download.folder}{The output folder to store created SBGN-ML files.}
}
\value{
A vector of character strings. The path to the created SBGN-ML files.
}
\description{
This function can generate a SBGN-ML file of our pre-collected SBGN-ML files
}
\examples{
data('pathways.info')
data(sbgn.xmls)
input.sbgn = downloadSbgnFile(
                  pathway.id = pathways.info[1,'pathway.id'],
                  download.folder = './')
}
