% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.utilities.R
\name{highlightArcs}
\alias{highlightArcs}
\title{Highlight arcs by arc class}
\usage{
highlightArcs(class = "all", color = "black", line.width = 2, tip.size = 6)
}
\arguments{
\item{class}{A character string. Default: "arc". The arc class to modify.}

\item{color}{A character string. Default: "black". The color of arcs with selected 'class'.}

\item{line.width}{Numeric. Default: 2. Line thickness.}

\item{tip.size}{Numeric. Default: 6. Tip size.}
}
\value{
A SBGNview object
}
\description{
This function can modify a SBGNview object's arc. Normally we use it as an argument to \code{+.SBGNview} and modify a SBGNview object. Run \code{help("+.SBGNview")} for more information.
}
\examples{
\dontrun{
obj.new <- SBGNview.obj + highlightArcs(class = 'production', color = 'red') 
}
}
