% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapped.ids}
\alias{mapped.ids}
\title{IDs mappable by SBGNview}
\format{
A list with two elements: A vector of mappable gene ID types (\code{mapped.ids[["gene"]]}), a vector of mappable compound ID types (\code{mapped.ids[["cpd"]]}).
}
\usage{
data(mapped.ids)
}
\description{
IDs mappable by SBGNview
}
\details{
ID types mappable to glyph IDs in SBGNhub SBGN-ML file collection.
}
\keyword{datasets}
