% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGNview.obj.fun.R
\name{outputFile}
\alias{outputFile}
\title{Retrieve output file information from a SBGNview object}
\usage{
outputFile(obj)
}
\arguments{
\item{obj}{A SBGNview class object.}
}
\value{
A string. The output file information.
}
\description{
Retrieve output file information from a SBGNview object
}
\details{
This function prints the output file path recorded in a SBGNview object.
}
\examples{
\dontrun{
outputFile(SBGNview.obj) 
}
}
