% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways.info}
\alias{pathways.info}
\title{Information of collected pathways}
\format{
A data.frame
}
\usage{
data(pathways.info)
}
\description{
Information of collected pathways
}
\details{
The information of pre-collected pathways and their SBGN-ML file information, such as pathway ID, name, source database, glyph ID types and URLs to the original pathway webpages.
}
\keyword{datasets}
